+++ sdk/core/azure-core-test/inc/azure/core/test/test_base.hpp
@@ -100,7 +100,7 @@ namespace Azure { namespace Core { namespace Test {
 
     // Creates the sdk client for testing.
     // The client will be set for record and playback before it is created.
-    Azure::Core::Credentials::TokenCredentialOptions GetTokenCredentialsOptions()
+    Azure::Core::Credentials::TokenCredentialOptions GetTokenCredentialOptions()
     {
       // Run instrumentation before creating the client
       Azure::Core::Credentials::TokenCredentialOptions options;
